requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
boto3==1.28.31
botocore==1.31.31
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
idna==3.4
jmespath==1.0.1
kombu==5.3.2
pip-autoremove==0.10.0
pipreqs==0.4.13
prompt-toolkit==3.0.39
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
yarg==0.1.9
